+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 + Added **Plot Price** - Show the price of the plot in coins when inside the Configure Plots inventory.
 + Added **Garden Crop Milestone Display** - Shows the progress and ETA until the next crop milestone is reached and the current crops/minute value. (Requires a tool with either a counter or cultivating enchantment)
 + Added **Best Crop Display** - Lists all crops and their ETA till next milestone. Sorts for best crop for getting garden level or skyblock level.
+* Added **Copper Price** - Show the price for copper inside the visitor gui.
 
 
 ### Features from other Mods